backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
asyncpg==0.29.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.3
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# External APIs
requests==2.31.0
aiohttp==3.9.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Cache
redis==5.0.1
aiocache==0.12.2

# Utils
python-dateutil==2.8.2
pytz==2023.3